+++ src/components/ContentWeb/index.tsx
@@ -1,0 +1,6 @@
+import { PropsWithChildren } from 'react'
+import { Container } from './styles'
+
+export function ContentWeb({ children }: PropsWithChildren) {
+  return <Container>{children}</Container>
+}
