--- src/components/ProductWeb/styles.ts
@@ -1,50 +1,0 @@
-import { styled } from '@/styles/index'
-
-export const ProductContainer = styled('main', {
-  display: 'grid',
-  gridTemplateColumns: '1fr 1fr',
-  alignItems: 'stretch',
-  gap: '4rem',
-})
-
-export const ImageContainer = styled('div', {
-  height: 600,
-  background: 'linear-gradient(180deg, $orange500 0%, $gray900 100%)',
-  borderRadius: 8,
-  padding: '0.25rem',
-
-  display: 'flex',
-  alignItems: 'center',
-  justifyContent: 'center',
-
-  img: {
-    objectFit: 'cover',
-  },
-
-  '@media (max-height: 700px)': {
-    height: 450,
-  },
-})
-
-export const ProductDetails = styled('div', {
-  display: 'flex',
-  flexDirection: 'column',
-
-  h1: {
-    fontSize: '$2xl',
-    color: '$white',
-  },
-
-  span: {
-    display: 'block',
-    fontSize: '$2xl',
-    color: '$orange500',
-  },
-
-  p: {
-    marginTop: '2.5rem',
-    fontSize: '$md',
-    lineHeight: 1.6,
-    color: '$gray300',
-  },
-})
