+++ src/storage/index.ts
@@ -1,20 +1,20 @@
+import { ProductInfoProps } from '@/types/product'
 import { legacy_createStore as createStore } from 'redux'
 
 import { persistStore, persistReducer } from 'redux-persist'
 import storage from 'redux-persist/lib/storage'
 
-import { productProps } from './modules/cart/action'
 import rootReducer from './modules/rootReducer'
 
 export interface reduxProps {
-  cart: productProps[]
-  products: productProps[]
+  cart: ProductInfoProps[]
+  products: ProductInfoProps[]
 }
 
 const persistConfig = {
   key: 'root',
   storage,
-  whitelist: ['cart'],
+  whitelist: ['cart', 'products'],
 }
 
 const persistedReducer = persistReducer(persistConfig, rootReducer)
