--- src/types/home.ts
@@ -1,11 +1,0 @@
-export interface ProductInfoProps {
-  id: string
-  name: string
-  imageUrl: string
-  price: string
-  unitLabel: string
-}
-
-export interface HomeProps {
-  products: ProductInfoProps[]
-}
