+++ src/types/product.ts
@@ -1,11 +1,18 @@
+export interface ProductInfoProps {
+  id: string
+  name: string
+  imageUrl: string
+  price: string
+  description: string
+  defaultPriceId: string
+  quantity: number
+  unitLabel: string
+}
+
+export interface ProductsProps {
+  products: ProductInfoProps[]
+}
+
 export interface ProductProps {
-  product: {
-    id: string
-    name: string
-    imageUrl: string
-    price: string
-    description: string
-    defaultPriceId: string
-    quantity: number
-  }
+  product: ProductInfoProps
 }
